+++ server/base/StringWithReplacement.cpp
@@ -50,7 +50,7 @@ void StringWithReplacement::set(const std::string &query)
             /* [0]: occurence to replace
              * [1,2]: total size of the String
              * List of: 16Bit header + string content */
-            const uint16_t &arraysize=1+2+(numberOfReplace+1/*if one %1 mean 2 string, if  %1,%2 mean 3 string*/)*sizeof(uint16_t)+query.size();
+            const uint16_t &arraysize=static_cast<uint16_t>(1+2+(numberOfReplace+1/*if one %1 mean 2 string, if  %1,%2 mean 3 string*/)*sizeof(uint16_t)+query.size());
             char preparedQueryTemp[arraysize];
             preparedQueryTemp[0]=numberOfReplace;
             preparedQueryTemp[1]=0;
@@ -72,7 +72,7 @@ void StringWithReplacement::set(const std::string &query)
                         #endif
                         return;
                     }
-                    const uint16_t &size=foundinternal-previousStringPos;
+                    const uint16_t &size=static_cast<uint16_t>(foundinternal-previousStringPos);
                     *reinterpret_cast<uint16_t *>(preparedQueryTemp+pos)=size;
                     pos+=2;
                     if(size>0)
@@ -81,7 +81,7 @@ void StringWithReplacement::set(const std::string &query)
                         memcpy(preparedQueryTemp+pos,extractedPart.data(),size);
                         pos+=size;
                     }
-                    previousStringPos=foundinternal+testToFind.size();
+                    previousStringPos=static_cast<uint16_t>(foundinternal+testToFind.size());
                 }
                 else
                 {
@@ -102,7 +102,7 @@ void StringWithReplacement::set(const std::string &query)
                 return;
             }
             //the last part:
-            const uint16_t &size=query.size()-previousStringPos;
+            const uint16_t &size=static_cast<uint16_t>(query.size()-previousStringPos);
             *reinterpret_cast<uint16_t *>(preparedQueryTemp+pos)=size;
             pos+=2;
             if(size>0)
